requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
AMFM_decompy==1.0.11
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Columnar==1.4.1
comm==0.2.2
contourpy==1.3.1
cupy-cuda12x==13.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastrlock==0.8.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiwer==3.0.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.9.2
lightning==2.4.0
lightning-utilities==0.11.8
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pesq==0.0.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
progressbar2==4.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.2.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytorch-lightning==2.4.0
pytz==2024.2
pyvad==0.2.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
stack-data==0.6.3
statsmodels==0.14.5
string-color==1.2.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torcheval==0.0.7
torchmetrics==1.6.0
torchpq==0.3.0.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtual-dataset==1.0.1
warpq==1.5.2
wcwidth==0.2.13
webrtcvad==2.0.10
Werkzeug==3.1.3
yarl==1.17.1
zerospeech-libriabx==1.0.5
